Move cron functions to job package
