--- server/state.go
@@ -1,7 +1,0 @@
-package server
-
-import "github.com/harryzcy/snuuze/platform"
-
-type State struct {
-	Repos []platform.Repo
-}
